requirements.txt
openai-agents==0.0.5
openai>=0.27.0
PyYAML>=6.0
python-dotenv>=1.0.0
pytest>=7.0
